ComfyUI-Addoor/requirements.txt
torch>=1.7.0
numpy>=1.19.0
Pillow>=8.0.0

